+++ crates/proof-of-sql/src/base/scalar/scalar.rs
